+++ src/main/java/com/kaspi/payment/dto/PaymentResponseDto.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 public class PaymentResponseDto {
@@ -11,17 +13,17 @@ public class PaymentResponseDto {
     private String message;
     private Long paymentId;
     private Long dealId;
-    private Integer amount;
+    private BigDecimal amount;
     private String status; // "MATCHED", "NO_MATCH", "ERROR", "DUPLICATE"
 
     // Factory методы для удобства
-    public static PaymentResponseDto matched(Long paymentId, Long dealId, Integer amount) {
+    public static PaymentResponseDto matched(Long paymentId, Long dealId, BigDecimal amount) {
         return new PaymentResponseDto(
                 true, "Payment matched with deal",
                 paymentId, dealId, amount, "MATCHED");
     }
 
-    public static PaymentResponseDto noMatch(Long paymentId, Integer amount) {
+    public static PaymentResponseDto noMatch(Long paymentId, BigDecimal amount) {
         return new PaymentResponseDto(
                 true, "No matching deal found",
                 paymentId, null, amount, "NO_MATCH");
@@ -38,7 +40,7 @@ public class PaymentResponseDto {
                 false, error, null, null, null, "ERROR");
     }
 
-    public PaymentResponseDto(boolean success, String message, Long paymentId, Long dealId, Integer amount, String status) {
+    public PaymentResponseDto(boolean success, String message, Long paymentId, Long dealId, BigDecimal amount, String status) {
         this.success = success;
         this.message = message;
         this.paymentId = paymentId;
